fixing arg types on externals
git-svn-id: https://jamoma.svn.sourceforge.net/svnroot/jamoma/trunk@134 464221e2-0a0f-0410-983b-f8d57d210c17

+++ Jamoma/externals/jmod_change/jmod.change.cpp
@@ -122,7 +122,7 @@ void change_int(t_change *x, long value)
 	x->last_input_int = value;		
 }
 
-void change_float(t_change *x, float value)
+void change_float(t_change *x, double value)
 {
 	if(x->inletnum == 0){
 		if(value != x->last_input_float)
